Switch_cases
So today i learned about the switch case statements basically the idea is that when you want to execute a statement when user enters a specific value then comes into picture
switch cases. They are statements or a block of code if someone wants to print or display a statement like say when we want to number of days in a week we can do it by switch cases. I've tried in the above mentioned file